+++ lib/src/generator/templates.aot_renderers_for_html.dart
@@ -1048,10 +1048,31 @@ String renderIndex(PackageTemplateData context0) {
           <h3>''');
       buffer.writeEscaped(context8.name);
       buffer.write('''</h3>''');
-      var context9 = context8.publicLibrariesSorted;
+      var context9 = context8.externalItems;
       for (var context10 in context9) {
+        buffer.writeln();
+        buffer.write('''
+            <dt>
+              <span class="name">
+                <a href="''');
+        buffer.writeEscaped(context10.url);
+        buffer.write('''" target="_blank">
+                  ''');
+        buffer.writeEscaped(context10.name);
+        buffer.writeln();
+        buffer.write('''
+                  <span class="material-symbols-outlined">open_in_new</span>
+                </a>
+              </span>
+            </dt>
+            <dd>''');
+        buffer.writeEscaped(context10.docs);
+        buffer.write('''</dd>''');
+      }
+      var context11 = context8.publicLibrariesSorted;
+      for (var context12 in context11) {
         buffer.write('\n            ');
-        buffer.write(_renderIndex_partial_library_2(context10));
+        buffer.write(_renderIndex_partial_library_2(context12));
       }
     }
     buffer.writeln();
@@ -4336,12 +4357,28 @@ String _deduplicated_lib_templates_html__packages_html(
       <li class="section-subtitle">''');
       buffer.writeEscaped(context9.name);
       buffer.write('''</li>''');
-      var context10 = context9.publicLibrariesSorted;
+      var context10 = context9.externalItems;
       for (var context11 in context10) {
         buffer.writeln();
         buffer.write('''
+        <li class="section-subitem">
+          <a href="''');
+        buffer.writeEscaped(context11.url);
+        buffer.write('''" target="_blank">
+            ''');
+        buffer.writeEscaped(context11.name);
+        buffer.writeln();
+        buffer.write('''
+            <span class="material-symbols-outlined">open_in_new</span>
+          </a>
+        </li>''');
+      }
+      var context12 = context9.publicLibrariesSorted;
+      for (var context13 in context12) {
+        buffer.writeln();
+        buffer.write('''
         <li class="section-subitem">''');
-        buffer.write(context11.linkedName);
+        buffer.write(context13.linkedName);
         buffer.write('''</li>''');
       }
     }