+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Remove explicit library names. (#3609)
 * No longer write the dartdoc version into generated html files.
+* Add support for GitHub markdown alert syntax
+  (https://docs.github.com/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#alerts).
 
 ## 8.0.3