+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove explicit library names. (#3609)
 * No longer write the dartdoc version into generated html files.
 * Add support for [GitHub markdown alert syntax][].
+* Add support for external links to dartdoc categories (#3617).
 * Require Dart 3.2.0 or later.
 * Require `analyzer: ^6.4.0`.